+++ apps/transifex/src/locations/Sidebar.jsx
@@ -43,7 +43,7 @@ function Sidebar() {
     {
       type: 'plain',
       defaults: {
-        environmentId: sdk.ids.environmentAlias ?? sdk.ids.environment,
+        environmentId: sdk.ids.environment,
         spaceId: sdk.ids.space,
       },
     },